+++ Project_SWP391_OnlineShopping_Nhom3/src/java/controller/Admin/InsertProduct.java
@@ -92,7 +92,7 @@ public class InsertProduct extends BasedRequiredAuthenticationController1 {
         }else
             status = false;
         dao.insertProduct(pid, name, img, num, price, description, status, cid);
-        response.sendRedirect(request.getContextPath() + "/AdminConsole");
+        response.sendRedirect(request.getContextPath() + "/AdminProductList");
     }
 
     /**
