+++ Project_SWP391_OnlineShopping_Nhom3/src/java/controller/Admin/RemoveProduct.java
@@ -29,22 +29,6 @@ public class RemoveProduct extends BasedRequiredAuthenticationController1 {
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        try (PrintWriter out = response.getWriter()) {
-            /* TODO output your page here. You may use following sample code. */
-            out.println("<!DOCTYPE html>");
-            out.println("<html>");
-            out.println("<head>");
-            out.println("<title>Servlet RemoveProduct</title>");            
-            out.println("</head>");
-            out.println("<body>");
-            out.println("<h1>Servlet RemoveProduct at " + request.getContextPath() + "</h1>");
-            out.println("</body>");
-            out.println("</html>");
-        }
-    }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     /**
@@ -61,7 +45,7 @@ public class RemoveProduct extends BasedRequiredAuthenticationController1 {
         String pid = request.getParameter("pid");
         DAO dao = new DAO();
         dao.removeProduct(pid);
-        response.sendRedirect(request.getContextPath() + "/AdminConsole");
+        response.sendRedirect(request.getContextPath() + "/AdminProductList");
     }
 
     /**
@@ -75,7 +59,6 @@ public class RemoveProduct extends BasedRequiredAuthenticationController1 {
     @Override
     protected void processPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        processRequest(request, response);
     }
 
     /**
