+++ Project_SWP391_OnlineShopping_Nhom3/src/java/controller/Admin/UpdateProduct.java
@@ -32,22 +32,7 @@ public class UpdateProduct extends BasedRequiredAuthenticationController1 {
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        try (PrintWriter out = response.getWriter()) {
-            /* TODO output your page here. You may use following sample code. */
-            out.println("<!DOCTYPE html>");
-            out.println("<html>");
-            out.println("<head>");
-            out.println("<title>Servlet UpdateProduct</title>");            
-            out.println("</head>");
-            out.println("<body>");
-            out.println("<h1>Servlet UpdateProduct at " + request.getContextPath() + "</h1>");
-            out.println("</body>");
-            out.println("</html>");
-        }
-    }
+    
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     /**
@@ -83,7 +68,7 @@ public class UpdateProduct extends BasedRequiredAuthenticationController1 {
     @Override
     protected void processPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String pid = request.getParameter("ppid");
+        String pid = request.getParameter("pid");
         String pname = request.getParameter("name");
         int pquantity = Integer.parseInt(request.getParameter("quantity"));
         int pprice = Integer.parseInt(request.getParameter("money"));
@@ -100,7 +85,7 @@ public class UpdateProduct extends BasedRequiredAuthenticationController1 {
 
         DAO dao = new DAO();
         dao.updateProduct(pid, pname, pquantity, pprice, pimage, pdescription, status, CateID);
-        response.sendRedirect(request.getContextPath() + "/AdminConsole");
+        response.sendRedirect(request.getContextPath() + "/AdminProductList");
     }
 
     /**
