+++ Project_SWP391_OnlineShopping_Nhom3/web/WEB-INF/web.xml
@@ -90,7 +90,7 @@
     </servlet>
     <servlet>
         <servlet-name>AdminProductList</servlet-name>
-        <servlet-class>controller.AdminProductList</servlet-class>
+        <servlet-class>controller.Admin.AdminProductList</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>LoginController</servlet-name>
